[SYCL][E2E] Add comment with issue for inline ASM leaks (#15130)
The tests in sycl/test-e2e/InlineAsm/Negative are disabled due to
reported module leaks. The reason for this "leak" is because the modules
fail to build, and as such are never actually created. The false
positive has been reported in
https://github.com/oneapi-src/unified-runtime/issues/1990. This commit
adds a link to this issue in the affected tests.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>