[Loader] Fix xpti finalization order
Global xptiContext was beeing (sometimes) destroyed before the
tracing context (as both were global variables, their order of
destruction was not guaranteed). This resulted in lost
notifications as xptiFinalize was being called after xptiFrameworkFinalize.

Fix this by creating a global shared_ptr to the xptiContext.
Bump the refcount of that pointer every time a context is created
and decrease when it is destroyed to make sure that xptiFrameworkInitialize()
is not beeing called before the last context is destroyed.