Fix subset devicelib linking for NVPTX
Ensure that when subsets of devicelibs are excluded or included
by fno-sycl-device-lib or fsycl-device-lib, the correct libraries
are linked.
For -fsycl-device-lib this means that regardless of which libraries
are specified, the full single devicelib for NVPTX will be used.
For -fno-sycl-device-lib it means that its values will be ignored,
unless it contains "all", in which case only internal libraries
will be linked against.

Fix an error when printing a warning that some device libraries
have been ignored for -fno-sycl-device-lib