[SYCL][Graph] Update testing of unsupported features (#15202)
Update testing for unsupported usage of sycl::reduction and sycl::stream
features in graph nodes.

* Remove E2E tests marked "NOT_YET_IMPLEMENTED". These are always
disabled by lit, and as a result never compiled. Resulting in the tests
silently breaking. We can bring these tests back from git history once
the features are implemented.
* Add unitest for using sycl streams in a graph node, also improve
implementation so that an exception is thrown when the node is added
informing the user that this is unsupported.