[SYCL] Fix windows exp(complex) edge case (#15162)
There are three test cases that added in
https://github.com/intel/llvm/pull/14329 that do not match the expected
result on Windows. This is due to the fact the supporting the std math
functions on device Windows is partially determined by the MSVC
implementation of the math functions.

- `tanh(-inf + nan * i)` and `tanh(-inf + -inf * i)` should both be `1 +
0*i`, but the MSVC implementation does not obey this equation even on
host. This PR changes the test to ignore these cases on windows.
- `exp(1e6 + -0.1*i)` should be `inf + -inf*i`, but was `inf + inf*i`.
This equation holds for MSVC implementation on host. Indeed this was
problem with an implementation of the `_FExp` helper function that
`exp(std::complex)` is implemented with. This is fixed in this PR.

Additionally, this PR changes some of DeviceLib tests were changed from
being unsupported on GPU to just be unsupported on HIP or CUDA.