[SYCL][XPTI] Refactoring framework to use 128-bit keys for collision elimination (#14467)
Previous implementation of the XPTI framework used 64-bit hash values to
represent trace points in the code and this has led to a few of hash
collisions.This refactoring moves to a 128-bit key to guarantee
uniqueness. The changes needed to SYCL runtime to fully migrate to newer
APIs will be pushed as a **separate Part 2 pull request**. Current pull
request include changes to the XPTI framework and minor changes to SYCL
runtime to reflect the transition to 128-bit keys and ensure validity of
the tests.
 - 128-bit keys for internal storage and lookups
 - Support 64-bit universal IDs for backward compatibility
 - Updated tests to handle legacy API and new APIs for correctness tests
- Updated performance tests to report metrics for both 64-bit and 28-bit
native APIs
- Updated SYCL instrumentation to return a new trace event for each
instance of a trace point. Earlier implementation always returned the
same trace event for a give trace point as the metadata associated with
a trace event was deemed to be invariant. However, with the need for
mutable metadata, this change is required.
 - Minor updates to documentation

**NOTE**: Since more events are generated due to the creation of a new
trace event for each trace point instance, some tests that rely on event
sequences may have to be updated.

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>
Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>
Co-authored-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>
Co-authored-by: Artur Gainullin <artur.gainullin@intel.com>