[SYCL] Add host kernel instantiation for debuggers
The changes in https://github.com/intel/llvm/pull/14460 removed the
seemingly unused functions for running kernels on on host. However, this
turned out to be used by debuggers as they need the kernel code to be
in the host executable.

This commit adds a simplified version of the kernel instantiation that
the aforementioned patch removed.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>