[Driver][SYCL] Use existing option for device-only and some cleanup
Update -fsycl-device-only usage to be an alias to --offload-device-only.
This requires an additional unique check for SYCL enabling due to the
expectation that -fsycl-device-only also implies -fsycl.

Update associated tests that use -fsycl-device-only -emit-llvm.  Our
default is already bitcode output, fixing this up aligns better with
community expectations when using --offload-device-only.

Also do some variable cleanup to better be in sync with community
variable names. And use a more common diagnostic for invalid options
with -fsycl.