[SYCL] Limit allocation size in atomic_memory_order_seq_cst test to a safe maximum size (#14195)
The `atomic_memory_order_seq_cst.cpp` test only checks
`info::device::max_mem_alloc_size` to limit the maximum allocation. It
does not check the actual available on device. This lead to allocating
exceeding amount of memory that I caught while running the test on
Nvidia GTX 1650 sm_75, because the test does not check the amount of
memory on the device that is free according to the OS. This commit makes
it consider `ext::intel::info::device::free_memory` when selecting the
number of iterations to perform. While that is still an _intel_
extension, it is currently implemented on the relevant backend targets
for this test.