[SYCL][E2E LIT] Remove append_path=True from config.extra_environment handling (#15279)
When append_path is True, value is normalized to lowercase on Windows at

https://github.com/llvm-mirror/llvm/blob/2c4ca6832fa6b306ee6a7010bfb80a3f2596f824/utils/lit/lit/llvm/config.py#L139
For env value that isn't a path, the normalization is unexpected.

For env value that is a path, it doesn't make sense to append_path
unconditionally. Test driver should append path when necessary before
setting a env.