[SYCL][NATIVECPU] Fix local scope module variables for native cpu (#15280)
Although local scope variables inside the kernel are less common in
SYCL, they can happen with hierarchical.

This fixes the problem by adding a pass to replace the local scope
variables which start life as globals with a struct which is allocated
on the stack.

Additionally, this required updating of the code which renames and
removes kernel based on wrappers and vecz success. To simplify this we
run the OCK utility pass TransferKernelMetadata which adds metadata to
store the original kernel name. This in turn simplifies this code
significantly.

Note this fixes fails in kernel/kernel_attributes_wg_hint.cpp SYCL CTS
for native cpu, which is being tested locally.