[SYCL] Fix handling of interop events for barrier with waitlist (#15352)
Currently Command::getUrEventsBlocking is responsible for preparing a
waitlist of UR events for the barrier.
This method used wrong assumption that if isEnqueued() returns false for
the event then it doesn't have UR handle because it was not enqueued. So
if there is an associated command we would enqueue it to get the desired
UR handle, or we would just ignore this event if there is no associated
command.

Problem is that sycl::event created with interoperability constructor
has isEnqueued() as false (as it is not enqueued by SYCL RT) but it has
UR handle provided by user.
Before this patch we just ignored such event as it doesn't have
associated command and we didn't put it to the resulting list.

This patch fixes this problem by handling interop events properly in
this code path.