[SYCL] Fix devicelib identification for NVPTX
Recent upstream changes removed the workaround that required libraries
to be named ".cubin".

We also renamed the NVPTX libdevice library when we unified it into a
single bytecode library.

Unfortunately both of these things broke the driver's brittle logic that
determined whether or not an object was a device library. This resulted
in us not stripping any symbols from the final object, which is both
slower and broke some SYCL CTS tests.

This commit more clearly delineates logic between NVPTX and Native-CPU,
as the latter target didn't actually link against anything containing
'libdevice' so was too lenient.