[SYCL][E2E] Fix warnings from using GNU style options with MSVC compiler driver (#15364)
Due to the addition of the `-Werror` flag, these warnings were causing
test failures when compiling tests with clang-cl. This patch does the
following:

- Add `-Wno-unused-command-line-argument` to linking run lines. This is
needed because when testing the MSVC driver the `/EHsc` flag is added to
all run lines that call the compiler. This flag however is only needed
at the compile stage so it is reported as unused if a run line is only
linking.
- Add new expansion for no optimizations `%no_opt`, which is either
`-O0` or `/Od` depending on the compiler driver.
- Use the expansion `%cxx_std_option` in place of `-std=`
- For flags that do not necessarily have an MSVC equivalent, prepend
`/clang:` to the flag when using MSVC driver.