[SYCL][Driver] Update support for HIP/AMDGCN targets in DeviceConfigFile
This change adds new supported targets, gfx941 and gfx942.
It also fixes the supported subgroup size config value for AMD RDNA GPUs, due
to the fact that the ROCm driver does not allow support for wave64 mode in HIP,
for gfx10 and gx11 family of GPUs, which are based on the RDNA architecture.
This is a result of wavefront/cross-lane functions that only work with wave32.