[SYCL] Fix in-order queue dependencies for no scheduler path (#15412)
Runtime could submit kernel directly to scheduler if no buffers/streams
are used and if event dependencies are already handled by queue (in case
if it is in-order one). Although check if dependencies are submitted to
the same queue was missed. Now we add events submitted to another queue
but on the same context to event list in kernel launching.

---------

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>