[SYCL] Call clang instead of llc. (#15259)
Previously, OffloadWrapper::ConstructJob would call llc to convert LLVM
IR into machine code. With this change, it calls clang instead.

This should be mostly NFC for x86, there may be some small changes in
exactly which passes run, but they are intended to be roughly the same,
and enables future changes passing along more clang options.

For other targets, this is a bugfix. Specifically, for RISC-V, llc and
clang disagree on which subtarget features to enable by default,
resulting in linker errors due to subtarget feature mismatches between
the llc-compiled wrapper and other clang-compiled object files.