[SYCL][Ext][Bindless] Initial implementation of image spirv builtins on HIP
This PR implements the required spirv builtins in the libclc for AMD targets
to support image fetching (excluding sampled fetch for now) and sampling
as well as image array fetching and sampling.

Additionally, End-to-end tests are updated to require the aspects corresponding
to the feature that is being tested from the Bindless Images extension.
This helps avoid having to manually say which backend adapter is supported or
unsupported and instead rely on support based on aspect/device queries to drive
the execution of the tests.

Signed-off-by: Georgi Mirazchiyski <georgi.mirazchiyski@codeplay.com>