Don't use empty match test files
This patch removes any empty ".match" files, and instructs ctest to
run the tests verbatim if there is no match file available.

As well as providing more useful output, this also handles the case
where the test startup and/or teardown causes the test runner to fail.
Previously, this would be ignored (as it produced no output).

I also modified cts_exe.py so that it uses the same parameters as
ctest, so the same command invocation could be used for both.