[SYCL][E2E] Enable bfloat16_example test everywehre
The test was recently refactored and got accidentally enabled on HIP,
causing post-commit failure. It failed because it had been originally
written to only support JIT targets.

There seems to be no good reason for this test not to be run on CUDA &
HIP backends, so instead of restricting devices it should be run on this
patch refactors it to support AOT targets as well.