[SYCL] Fix a thread pool data race during shutdown
The flag that kills the threads waiting for work was an atomic bool,
but that is not enough: any modification of a condition must be made
under the mutex associated with that condition variable. Otherwise,
the bool value flip and the call to notify might happen after the
condition check in the other thread, but before it starts waiting for
notifications.