[SYCL][Graph] Reenable L0 interop test on PVC (#15493)
Re-enable the interop-level-zero-launch-kernel.cpp E2E test on PVC by
making the following correctness changes to the native L0 code:

* Both `zeKernelSetGroupSize` and `ZeGroupCount` were set to have 1024
in the X dimensions when this is the total size of the USM allocation,
and therefore the number of work items we want. Updated to use
`zeKernelSuggestGroupSize` to work calculate the correct group sizes and
count.
* Update `zeKernelSetArgumentValue` call so that the size parameter is
the size of the pointer rather than size of the USM allocation.