Split source checks into a seperate workflow
The "Build - Ubuntu" job has a stage that checks the source generated
from the spec matches the version stored in the repo. This is the
part of the job that takes the longest, and is ran multiple times.

Since it doesn't require a build, it has been brought out into its own
workflow which is only ran once. This should hopefully speed up builds.