[SYCL][clang] Always lift SYCL device code restrictions in constexpr contexts
Earlier a mechanism to lift SYCL device code restrictions in C++
manifestly constant-evaluated expressions was implemented in the front-end.
It was enabled under an option since it wasn't specified.
Since then it has been decided to make this behavior default for SYCL 2020.
This PR removes the option that was enabling the new behavior and
enables it by default.