[Driver][SYCL] Bound architecture mismatch with multiple targets (#15501)
When passing -fsycl-targets to specify targets to offload to, the user
can pass multiple targets. When those targets are meant to set various
associated architectures, those architectures should only be set for a
singular given toolchain/target.

Specifying a setting like -fsycl-targets=nvptx64,spir64_gen was setting
the wrong device architecture for the spir64_gen compilation. The
associated boundarch for nvptx64 (in this case sm_50) was being pushed
to the spir64_gen target, causing the wrong device value to be used for
the AOT compilation.

Fix this issue by fixing the logic in which we were assigning the bound
architecture to a given triple. The logic was not taking into account
non-spir64_gen targets, assigning the wrong arch when spir64_gen was
encountered after the nvptx64 target.