[SYCL][NATIVE_CPU] Fill in any SYCL functions which require mapping to mux (#15592)
Native cpu can make calls to mux builtins such as shuffle which are ABI
compliant but are not what is expected by ock passes. This fixes them up
by remove the vector versions from libnativecpu.cpp and using a pass to
convert from parameters which relate to the ABI to calling the mux
functions with the set interface unaffected by the ABI.

This currently only handle a small number of cases for shuffle such as
when a vector i2 is replaced with double or byval is used. It will be
expanded over time as needed.