[SYCL][Graph] Fix barrier transitive queue recording
It has been discovered that when a graph event is passed as a dependency
to a barrier operation on another in-order queue. This does not trigger
the queue to record the operation as it should. Fix this by not using
optimized path for in-order queues when graph event is a dependency,
with regression test for fix.