[SYCL][E2E] Track amount of XFAILed tests
We have a problem with our approach to `XFAIL`ing failed tests - we
don't always submit trackers for those failures, thus simply hiding them
from us.

Some of tests we `XFAIL`ed have been in that state for years and without
a tracker submitted about them, we don't even know that there are
issues.

This patch introduces a new requirement for marking test as expected to
fail: every `XFAIL` directive has to be followed by `XFAIL-TRACKER`
which points to intel/llvm issue submitted to analyze the failure and
remove `XFAIL` directive.

The tracker can be referenced either by URL, or through GH shortcuts
`owner/repo#NNNNN`.

To ensure that the new requirement is followed, a new test was added
which checks amount of improper `XFAIL` directives (i.e. those which are
not followed by `XFAIL-TRACKER` directive).

Similar approach is expected to be applied later for `UNSUPPORTED`
directives, but it will be done as a separate PR.