[SYCL][NFCI] Refactor `unittest::UrArray`
This is a first patch in series of detaching unit-test classes from UR
names.

In a previous attempts to do so (#14815) there were concerns about what
to do with `UrArray` and this PR is another attempt (see also #15014) to
do so.

This PR renames `UrArray` into `LifetimeExtender` to better communicate
its purpose: data structures emitted by the compiler (and therefore used
by the runtime) to describe device image and their properties to not
store data, but only hold pointers to them. Therefore when we mock those
in our unit-tests we need to ensure that their lifetime is long enough
to cover the whole test.

This is a non-functional change by its spirit, but what used to be
`UrArray` is now hidden from writers of unit-tests and the interface is
switched to `std::vector` - that is done to hide an implementaiton
detail and simplify amount of knowledge required to write unit-tests.