[SYCL] Correctly spell out SM version macro when AOT compiling
There was a typo resulting in the compiler using `*_SM[0-9][9-9]__`
format when the SM was not supported:
`#define __SYCL_TARGET_NVIDIA_GPU_SM50__ 0`
vs:
`#define __SYCL_TARGET_NVIDIA_GPU_SM_50__ 1`
for the supported SM version.

Prefer the underscore (`*_SM_[0-9][9-9]__`), as this is what CUDA uses,
for example when specifying `gpu-name` switch for `ptxas`.