[SYCL] Remove UR objects release checks on Windows (#15425)
These e2e tests started failing after PI removal and replacing it with
UR ([PR](https://github.com/intel/llvm/pull/14145)) However, they were
not related to it.

On Windows, dlls unloading is inconsistent and if we try to release
these UR objects manually, inconsistent hangs happen due to a race
between unloading the UR adapters dlls (in addition to their dependency
dlls) and the releasing of these UR objects (The proxy loader have
solved this problem partially
[here](https://github.com/intel/llvm/pull/15262)). So, we currently
shutdown without releasing them and windows should handle the memory
cleanup.

This behaviour is the same old behaviour as before removing PI as on
Investigations on this. This was only hidden before PI removal as it was
calling the PI entry-point but doesn't make it to UR entry-point and
Filecheck logs check for objects release would pass as it only check for
the call to the PI entry-point without check that the call was a
successful call. That was the PI call for `piContextRelease` before PI
removal:

```
---> piContextRelease(
        <unknown> : 0000023CC0EBA6C0
) ---> API Called After Plugin Teardown, Functon Call ignored.
```

Fixes #14768
Fixes #14950
Fixes #14968