[Driver][SYCL] Host objects usage improvements with -fsycl-link=image (#15539)
When using -fintelfpga -fsycl-link=image, the host objects were packaged
as objects as they would a regular archive. This was not handled in the
same way as -fsycl-link=early, causing an issue with how the objects can
be read in, especially on Windows.

Update the -fsycl-link=image behavior when packaging the host objects to
match that of -fsycl-link=early. This allows for the host objects to be
unbundled and used as actual objects during the link. The Microsoft
linker does not expect 'main' to be in a library, so breaking this out
allows for proper linking with -fsycl-link=image archives.