[SYCL][Docs] Allow discarded events case for ext_oneapi_get_last_event
This commit changes the behavior for queue::ext_oneapi_get_last_event
when the queue either does not currently have a last event or the last
event was discarded. In these cases, the implementation will now insert
a marker event at the end of the queue which signifies the work related
to the last command in the queue.

Additionally, this allows us to relax the requirements for the extension
to now allow it to be used on queues where discarding of events is
enabled.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>