[SYCL] Removal of interfering assert  (#15612)
`find_matching_descendent_device` return type is `std::optional`,
meaning it does not need to return a match if none is found. The calling
code anticipates this. But the assert here precludes that path.
Removing.

This is causing a bug on one of our big systems so we need the fix. But
I'm not sure if a test needs to be added nor how to write such, as we
don't see this problem on the lab or CI systems.