[SYCL] Fix barrier submission after command which bypasses the scheduler
Currently we mistakenly don't mark commands like memcpy, copy, fill etc. as
enqueued, this happens because they are enqueued differently and bypass
the scheduler.

Problem is that if barrier is submitted depending on such event then it
is just omitted.

So, fix the problem by marking such commands as enqueued.