[SYCL][LIBCLC][NATIVECPU] Fix ldexp and popcount on Native CPU (#15687)
`ldexp` in `generic` (for `half`) and `popcount` in `native_cpu` are
implemented using clang builtins. This PR makes it so we call the
correct "overload" of the builtin:
* `__builtin_ldexpf16` for `half`
* `__builtin_popcountg` for unsigned integer types
* for signed integer types the implicit cast to `int` performed by
calling `__builtin_popcount` doesn't work due to sign extension, so we
explicitly cast to `unsigned char` or `unsigned short`, and use
`__builtin_popcountg`.